Add DEPENDS declarations to existing install scripts
Declare dependencies for all existing install scripts to enable
automatic dependency resolution. Also standardize script headers
with shebang and set -e where missing.

+++ install.sh.d/alacritty.sh
@@ -1,4 +1,5 @@
 #!/bin/bash -e
+# DEPENDS: cargo curl git stow wget
 
 VERSION=0.14.0
 
@@ -9,16 +10,13 @@ rm -f alacritty.tar.gz
 
 pushd alacritty-$VERSION
 
-# On installe le compilateur Rust
-curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
-
-# On s’assure qu’on a la dernière version
-rustup override set stable
-rustup update stable
-
 # On installe les dépendances
-sudo apt update
-sudo apt install --yes cmake pkg-config libfreetype6-dev libfontconfig1-dev libxcb-xfixes0-dev libxkbcommon-dev python3
+if [ "$(uname -s)" == "Linux" ]; then
+    if ! dpkg -s cmake pkg-config libfreetype6-dev libfontconfig1-dev libxcb-xfixes0-dev libxkbcommon-dev python3 >/dev/null ; then
+        sudo apt update
+        sudo apt install --yes cmake pkg-config libfreetype6-dev libfontconfig1-dev libxcb-xfixes0-dev libxkbcommon-dev python3
+    fi
+fi
 
 # On compile
 cargo build --release
@@ -26,8 +24,10 @@ cargo build --release
 # On installe
 sudo cp target/release/alacritty /usr/local/bin/alacritty
 sudo cp extra/logo/alacritty-term.svg /usr/share/pixmaps/Alacritty.svg
-sudo desktop-file-install extra/linux/Alacritty.desktop
-sudo update-desktop-database
+if [ "$(uname -s)" == "Linux" ]; then
+    sudo desktop-file-install extra/linux/Alacritty.desktop
+    sudo update-desktop-database
+fi
 
 popd
