+++ install.sh.d/bat.sh
@@ -1,8 +1,14 @@
+#!/bin/bash -e
+# DEPENDS: gdebi wget
+
 if [ "$(uname -s)" == "Darwin" ]; then
-    brew install bat
+    if ! brew ls --versions bat >/dev/null ; then
+        brew install bat
+    fi
 elif [ "$(uname -s)" == "Linux" ]; then
-    wget https://github.com/sharkdp/bat/releases/download/v0.12.1/bat_0.12.1_amd64.deb -O /tmp/bat_0.12.1_amd64.deb
-    sudo apt-get install gdebi
-    sudo gdebi -n /tmp/bat_0.12.1_amd64.deb
-    rm /tmp/bat_0.12.1_amd64.deb
+    if ! dpkg -s bat >/dev/null ; then
+        wget https://github.com/sharkdp/bat/releases/download/v0.12.1/bat_0.12.1_amd64.deb -O /tmp/bat_0.12.1_amd64.deb
+        sudo gdebi --non-interactive /tmp/bat_0.12.1_amd64.deb
+        rm /tmp/bat_0.12.1_amd64.deb
+    fi
 fi
