+++ install.sh.d/hub.sh
@@ -5,6 +5,10 @@ if [ "$(uname -s)" == "Darwin" ]; then
         brew install hub
     fi
 elif [ "$(uname -s)" == "Linux" ]; then
-    sudo apt-get install snapd
-    sudo snap install hub --classic
+    if ! dpkg -s snapd >/dev/null ; then
+        sudo apt-get install snapd
+    fi
+    if ! snap list hub >/dev/null ; then
+        sudo snap install hub --classic
+    fi
 fi
